requirements.txt
openai==0.27.7
pandas==2.0.1
pinecone_client==3.1.0
streamlit==1.22.0
beautifulsoup4==4.12.2
numpy==1.24.3